Move away from junit launcher tests
